+++ Sources/StreamChat/Query/Sorting/ChannelListSortingKey.swift
@@ -89,7 +89,3 @@ extension ChatChannel {
         lastMessageAt ?? createdAt
     }
 }
-
-// Remove when InferSendableFromCaptures is enabled
-// https://github.com/swiftlang/swift-evolution/blob/main/proposals/0418-inferring-sendable-for-methods.md
-extension PartialKeyPath: @retroactive @unchecked Sendable where Root == ChatChannel {}